+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed handling of node numbers in queries with multiple alternatives.
+
 ## [4.1.4] - 2026-04-18
 
 ### Fixed